requirements.txt
langchain
langchain_community
streamlit
PyPDF2
torch
sentence_transformers
weaviate-client